--- app/src/main/java/com/example/tellingme/data/di/RepositoryModule.kt
@@ -1,24 +1,0 @@
-package com.example.tellingme.data.di
-
-import com.example.tellingme.data.network.NetworkService
-import com.example.tellingme.data.repositoryimpl.LoginRepositoryImpl
-import com.example.tellingme.domain.repository.LoginRepository
-import dagger.Binds
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule {
-
-    @Binds
-    @Singleton
-    abstract fun provideLoginRepository(
-        loginRepositoryImpl: LoginRepositoryImpl
-    ): LoginRepository
-
-
-}
