--- app/src/main/java/com/example/tellingme/data/model/DefaultResponse.kt
@@ -1,5 +1,0 @@
-package com.example.tellingme.data.model
-
-data class DefaultResponse(
-    val name: String
-)
