--- app/src/main/java/com/example/tellingme/domain/repository/LoginRepository.kt
@@ -1,8 +1,0 @@
-package com.example.tellingme.domain.repository
-
-import com.example.tellingme.data.model.DefaultResponse
-import retrofit2.Response
-
-interface LoginRepository {
-    suspend fun kakaoLogin(): Response<DefaultResponse>
-}
