--- app/src/main/java/com/example/tellingme/presentation/viewmodel/LoginViewModel.kt
@@ -1,20 +1,0 @@
-package com.example.tellingme.presentation.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.tellingme.domain.usecase.LoginUseCase
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class LoginViewModel @Inject constructor(
-    private val loginUseCase: LoginUseCase
-): ViewModel() {
-
-    fun kakaoLogin() {
-        viewModelScope.launch {
-            val response = loginUseCase().body()
-        }
-    }
-}
