--- app/src/main/java/com/example/tellingme/util/TellingMeApplication.kt
@@ -1,9 +1,0 @@
-package com.example.tellingme.util
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TellingMeApplication: Application() {
-
-}
