+++ app/src/main/java/com/tellingus/tellingme/data/di/RepositoryModule.kt
@@ -1,0 +1,22 @@
+package com.tellingus.tellingme.data.di
+
+import com.tellingus.tellingme.data.repositoryimpl.LoginRepositoryImpl
+import com.tellingus.tellingme.domain.repository.LoginRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class RepositoryModule {
+
+    @Binds
+    @Singleton
+    abstract fun provideLoginRepository(
+        loginRepositoryImpl: LoginRepositoryImpl
+    ): LoginRepository
+
+
+}
