+++ app/src/main/java/com/tellingus/tellingme/data/model/DefaultResponse.kt
@@ -1,0 +1,5 @@
+package com.tellingus.tellingme.data.model
+
+data class DefaultResponse(
+    val name: String
+)
