+++ app/src/main/java/com/tellingus/tellingme/data/network/NetworkService.kt
@@ -1,0 +1,14 @@
+package com.tellingus.tellingme.data.network
+
+import com.tellingus.tellingme.data.model.DefaultResponse
+import retrofit2.Response
+import retrofit2.http.POST
+
+interface NetworkService {
+
+    @POST
+    suspend fun postKakaoLogin(
+
+    ): Response<DefaultResponse>
+
+}
