+++ app/src/main/java/com/tellingus/tellingme/domain/repository/LoginRepository.kt
@@ -1,0 +1,8 @@
+package com.tellingus.tellingme.domain.repository
+
+import com.tellingus.tellingme.data.model.DefaultResponse
+import retrofit2.Response
+
+interface LoginRepository {
+    suspend fun kakaoLogin(): Response<DefaultResponse>
+}
