+++ app/src/main/java/com/tellingus/tellingme/presentation/ui/feature/main/MainActivity.kt
@@ -1,0 +1,48 @@
+package com.tellingus.tellingme.presentation.ui.feature.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.tellingus.tellingme.presentation.ui.common.BUTTON_SIZE
+import com.tellingus.tellingme.presentation.ui.common.BUTTON_STATE
+import com.tellingus.tellingme.presentation.ui.common.PrimaryButton
+import com.tellingus.tellingme.presentation.ui.theme.TellingMeTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            TellingMeTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.background) {
+                    Greeting("Android")
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun Greeting(name: String, modifier: Modifier = Modifier) {
+    PrimaryButton(
+        state = BUTTON_STATE.DEFAULT,
+        size = BUTTON_SIZE.LARGE,
+        text = "버튼버튼"
+    ) {
+        
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    TellingMeTheme {
+        Greeting("Android")
+    }
+}
