+++ app/src/main/java/com/tellingus/tellingme/presentation/viewmodel/LoginViewModel.kt
@@ -1,0 +1,20 @@
+package com.tellingus.tellingme.presentation.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.tellingus.tellingme.domain.usecase.LoginUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+@HiltViewModel
+class LoginViewModel @Inject constructor(
+    private val loginUseCase: LoginUseCase
+): ViewModel() {
+
+    fun kakaoLogin() {
+        viewModelScope.launch {
+            val response = loginUseCase().body()
+        }
+    }
+}
