+++ app/src/main/java/com/tellingus/tellingme/util/TellingMeApplication.kt
@@ -1,0 +1,9 @@
+package com.tellingus.tellingme.util
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class TellingMeApplication: Application() {
+
+}
